refactor: LottoMachine 제거
- getPurchasableCount -> LottoStore.getPurchasableCount
- createRandomTicket -> LottoTicket.createByRandom
- createReceipt -> LottoStore.createReceipt
