+++ src/main/java/lotto/model/store/LottoReceipt.java
@@ -1,7 +1,5 @@
 package lotto.model.store;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.Objects;
 
 public class LottoReceipt {
