+++ src/main/java/lotto/model/store/LottoStore.java
@@ -1,44 +1,52 @@
 package lotto.model.store;
 
-import lotto.model.ticket.LottoNumber;
 import lotto.model.ticket.LottoTicket;
 
-import java.util.List;
 import java.util.Objects;
 
 public class LottoStore {
 
-    private final LottoMachine lottoMachine;
+    private final Money price;
 
-    public static LottoStore create(LottoMachine lottoMachine) {
-        if (Objects.isNull(lottoMachine)) {
-            throw new IllegalArgumentException("로또 머신이 없습니다.");
+    public static LottoStore create(Money price) {
+        if (Money.isNullOrZero(price)) {
+            throw new IllegalArgumentException("로또 가격은 0원일 수 없습니다.");
         }
 
-        return new LottoStore(lottoMachine);
+        return new LottoStore(price);
     }
 
-    public LottoStore(LottoMachine lottoMachine) {
-        this.lottoMachine = lottoMachine;
+    public LottoStore(Money price) {
+        this.price = price;
+    }
+
+    public long getPurchasableCount(Money money) {
+        if (Objects.isNull(money)) {
+            return 0L;
+        }
+
+        return (long) money.divide(price);
     }
 
     public PurchaseResult buyAutomatically(Money money) {
         long lottoCount = getPurchasableCount(money);
-        LottoTicket ticket = lottoMachine.createRandomTicket(lottoCount);
-        LottoReceipt receipt = lottoMachine.createReceipt(money, lottoCount);
+        LottoTicket ticket = LottoTicket.createByRandom(lottoCount);
+        LottoReceipt receipt = createReceipt(money, lottoCount);
 
         return PurchaseResult.of(ticket, receipt);
     }
 
-    public long getPurchasableCount(Money money) {
-        return lottoMachine.getPurchasableCount(money);
-    }
-
     public PurchaseResult buyManually(Money money, LottoTicket ticket) {
         long lottoCount = getPurchasableCount(money);
-        LottoTicket randomTicket = lottoMachine.createRandomTicket(lottoCount - ticket.count());
-        LottoReceipt receipt = lottoMachine.createReceipt(money, lottoCount);
+        LottoTicket randomTicket = LottoTicket.createByRandom(lottoCount - ticket.count());
+        LottoReceipt receipt = createReceipt(money, lottoCount);
 
         return PurchaseResult.of(ticket.append(randomTicket), receipt);
+    }
+
+    private LottoReceipt createReceipt(Money money, long count) {
+        Money totalPrice = price.multiply(count);
+
+        return LottoReceipt.from(money, totalPrice);
     }
 }
